# [1/3] Optimize Slime Update Weights: Remove QWen3MOE Load Weight Overhead