<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

- 1/3: https://github.com/sgl-project/sglang/pull/8751 （23s => 22s)
- 2/3: https://github.com/sgl-project/sglang/pull/8753  (22s => 12s)
- 3/3: https://github.com/sgl-project/sglang/pull/8754  (12s  => 10s)

This PR maps to the `ASYNC_EP8_optimization_1` which save 2s

<img width="767" height="608" alt="Screenshot 2025-08-03 at 11 38 03 PM" src="https://github.com/user-attachments/assets/e6095186-67be-42ef-9cd8-1f584161218f" />


<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

<!-- Describe the changes made in this PR. -->

## Accuracy Test


GSM Before this PR:
```bash
Accuracy: 0.935
Invalid: 0.000
Latency: 13.829 s
Output throughput: 1653.462 token/s
```


GSM After this PR:
```bash
Accuracy: 0.935
Invalid: 0.000
Latency: 13.648 s
Output throughput: 1692.839 token/s
```

<!-- If this PR affects model-side code (e.g., kernels, model architecture), please provide accuracy test results. Ref: https://docs.sglang.ai/references/accuracy_evaluation.html -->

## Benchmark & Profiling


<!-- If this PR is expected to impact performance, please provide benchmark and profiling results. Ref: https://docs.sglang.ai/references/benchmark_and_profiling.html -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.